+++ gradio/CHANGELOG.md
@@ -1,5 +1,11 @@
 # gradio
 
+## 5.40.0
+
+### Features
+
+- [#11092](https://github.com/gradio-app/gradio/pull/11092) [`c8c17f1`](https://github.com/gradio-app/gradio/commit/c8c17f1a909952616a01967839482153d1419b78) - Add gr.Dialogue component.  Thanks @freddyaboulton!
+
 ## 5.39.0
 
 ### Features